# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BizCode API
  version: 0.1.0
  description: |
    REST API for BizCode commercial management system.
    Runs as an Express 5 sidecar on localhost:3001.
    Responses use a JSON envelope `{ success: true, data: ... }` except `/api/health`.
    Session authentication is required for protected resources.
    Protected endpoints can optionally receive `x-bizcode-channel` to enforce channel scope authorization.
    All routes accept optional `X-Request-Id` (max 128 characters); invalid or empty values are ignored and a UUID is assigned. The effective id is echoed on every response (`X-Request-Id`; see runtime middleware `correlationId`).
    Interactive documentation (Swagger UI): http://localhost:3001/api-docs/ (same contract as this file).
  contact:
    name: BizCode Team
servers:
  - url: http://localhost:3001
    description: Local development / Tauri sidecar
...
paths:
...
  /api/facturas:
    parameters:
      - $ref: '#/components/parameters/BizcodeChannelHeader'
    get:
      tags: [facturas]
      summary: List invoices
      parameters:
        - $ref: '#/components/parameters/ListLimit'
        - $ref: '#/components/parameters/ListOffset'
      responses:
        '200':
          description: List of invoices with items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacturaListEnvelope'
        '500':
          $ref: '#/components/responses/ServerError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [facturas]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacturaInput'
      responses:
        '200':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacturaEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          description: Client suspended, insufficient stock, or other business rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
...
  /api/facturas/{id}/void:
    parameters:
      - $ref: '#/components/parameters/BizcodeChannelHeader'
    put:
      tags: [facturas]
      summary: Void an invoice
      description: |
        Sets `estado` to `N` (anulada), reverses the customer balance by the invoice total,
        and records an AuditEvent with the provided motivo. Requires `sales.cancel` permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoidInput'
      responses:
        '200':
          description: Invoice voided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacturaEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '409':
          description: Invoice already voided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
        '500':
          $ref: '#/components/responses/ServerError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: bizcode_session
  parameters:
    BizcodeChannelHeader:
      name: x-bizcode-channel
      in: header
      required: false
      description: Optional channel scope selector. When provided, it must be allowed by the authenticated user's `claims.scope.channels`.
      schema:
        type: string
        enum: [counter, field, backoffice, warehouse, delivery]
...
    ListLimit:
      name: limit
      in: query
      required: false
      description: Page size for list endpoints (default 100, max 500)
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
    ListOffset:
      name: offset
      in: query
      required: false
      description: Number of rows to skip before returning results
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    ApiErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
...
    ListPaginationMeta:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          minimum: 0
          description: Row count matching the list filter (before limit/offset)
        limit:
          type: integer
          minimum: 1
          maximum: 500
          description: Effective page size (same semantics as query `limit`)
        offset:
          type: integer
          minimum: 0
          description: Effective skip (same semantics as query `offset`)
...
    Articulo:
      type: object
      additionalProperties: true
      properties:
        id:
          type: integer
        codigo:
          type: integer
        descripcion:
          type: string
        rubroId:
          type: integer
        rubro:
          $ref: '#/components/schemas/Rubro'
        condIva:
          type: string
        umedida:
          type: string
        precioLista1:
          type: number
        precioLista2:
          type: number
        costo:
          type: number
        stock:
          type: integer
        minimo:
          type: integer
        activo:
          type: boolean
...
    Rubro:
      type: object
      additionalProperties: true
      properties:
        id:
          type: integer
        codigo:
          type: integer
        nombre:
          type: string
...
    Factura:
      type: object
      additionalProperties: true
      properties:
        id:
          type: integer
        fecha:
          type: string
          format: date-time
        tipo:
          type: string
        prefijo:
          type: string
        numero:
          type: integer
        clienteId:
          type: integer
        formaPagoId:
          type: integer
          nullable: true
        neto1:
          type: number
        neto2:
          type: number
        neto3:
          type: number
        iva1:
          type: number
        iva2:
          type: number
        total:
          type: number
        estado:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/FacturaItem'
    FacturaItem:
      type: object
      additionalProperties: true
      properties:
        id:
          type: integer
        articuloId:
          type: integer
        articulo:
          $ref: '#/components/schemas/Articulo'
        cantidad:
          type: number
        precio:
          type: number
        dscto:
          type: number
        subtotal:
          type: number
    VoidInput:
      type: object
      required: [motivo]
      properties:
        motivo:
          type: string
          minLength: 1
          maxLength: 500
          description: Reason for voiding the invoice. Stored in AuditEvent metadata.
    FacturaInput:
      type: object
      required: [fecha, tipo, numero, clienteId, neto1, neto2, neto3, iva1, iva2, total, items]
      properties:
        fecha:
          type: string
          format: date
        tipo:
          type: string
          enum: [A, B]
        prefijo:
          type: string
        numero:
          type: integer
        clienteId:
          type: integer
        formaPagoId:
          type: integer
          nullable: true
        neto1:
          type: number
        neto2:
          type: number
        neto3:
          type: number
        iva1:
          type: number
        iva2:
          type: number
        total:
          type: number
        items:
          type: array
          items:
            type: object
            required: [articuloId, cantidad, precio, dscto, subtotal]
            properties:
              articuloId:
                type: integer
              cantidad:
                type: number
              precio:
                type: number
              dscto:
                type: number
              subtotal:
                type: number
    FacturaListEnvelope:
      allOf:
        - type: object
          required: [success, data]
          properties:
            success:
              type: boolean
              const: true
            data:
              type: array
              items:
                $ref: '#/components/schemas/Factura'
        - $ref: '#/components/schemas/ListPaginationMeta'
    FacturaEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/Factura'
...
  responses:
    BadRequest:
      description: Request payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
    UnauthorizedError:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
    ForbiddenError:
      description: Authenticated but missing permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'